Move Api\AppendReservationsInterface and Api\ReservationBuilderInterface to Model namespace
