+++ app/code/Magento/InventoryReservations/Model/ReservationBuilder.php
@@ -11,8 +11,8 @@ use Magento\Framework\ObjectManagerInterface;
 use Magento\Framework\Validation\ValidationException;
 use Magento\Framework\Validation\ValidationResult;
 use Magento\Framework\Validation\ValidationResultFactory;
-use Magento\InventoryReservationsApi\Api\Data\ReservationInterface;
-use Magento\InventoryReservationsApi\Api\ReservationBuilderInterface;
+use Magento\InventoryReservationsApi\Model\ReservationInterface;
+use Magento\InventoryReservationsApi\Model\ReservationBuilderInterface;
 
 /**
  * @inheritdoc
