+++ app/code/Magento/InventoryReservations/Model/ResourceModel/SaveMultiple.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
 namespace Magento\InventoryReservations\Model\ResourceModel;
 
 use Magento\Framework\App\ResourceConnection;
-use Magento\InventoryReservationsApi\Api\Data\ReservationInterface;
+use Magento\InventoryReservationsApi\Model\ReservationInterface;
 
 /**
  * Implementation of Reservation save multiple operation for specific db layer
