+++ app/code/Magento/InventoryReservationsApi/Model/AppendReservationsInterface.php
@@ -5,9 +5,9 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventoryReservationsApi\Api;
+namespace Magento\InventoryReservationsApi\Model;
 
-use Magento\InventoryReservationsApi\Api\Data\ReservationInterface;
+use Magento\InventoryReservationsApi\Model\ReservationInterface;
 
 /**
  * Domain service used to append Reservations to keep track of quantity deductions before the related SourceItems
