+++ app/code/Magento/InventoryReservationsApi/Model/ReservationBuilderInterface.php
@@ -5,10 +5,10 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventoryReservationsApi\Api;
+namespace Magento\InventoryReservationsApi\Model;
 
 use Magento\Framework\Validation\ValidationException;
-use Magento\InventoryReservationsApi\Api\Data\ReservationInterface;
+use Magento\InventoryReservationsApi\Model\ReservationInterface;
 
 /**
  * Used to build ReservationInterface objects
