+++ app/code/Magento/InventorySales/Test/Integration/Stock/GetProductSalableQtyTest.php
@@ -8,8 +8,8 @@ declare(strict_types=1);
 namespace Magento\InventorySales\Test\Integration\Stock;
 
 use Magento\InventoryReservationsApi\Model\CleanupReservationsInterface;
-use Magento\InventoryReservationsApi\Api\AppendReservationsInterface;
-use Magento\InventoryReservationsApi\Api\ReservationBuilderInterface;
+use Magento\InventoryReservationsApi\Model\AppendReservationsInterface;
+use Magento\InventoryReservationsApi\Model\ReservationBuilderInterface;
 use Magento\InventorySalesApi\Api\GetProductSalableQtyInterface;
 use Magento\TestFramework\Helper\Bootstrap;
 use PHPUnit\Framework\TestCase;
