+++ app/code/Magento/InventorySales/Test/Integration/StockManagement/ReservationPlacingOnUseConfigManageStockSetToZeroTest.php
@@ -9,8 +9,8 @@ namespace Magento\InventorySales\Test\Integration\StockManagement;
 
 use Magento\InventoryReservationsApi\Model\CleanupReservationsInterface;
 use Magento\InventoryReservationsApi\Model\GetReservationsQuantityInterface;
-use Magento\InventoryReservationsApi\Api\AppendReservationsInterface;
-use Magento\InventoryReservationsApi\Api\ReservationBuilderInterface;
+use Magento\InventoryReservationsApi\Model\AppendReservationsInterface;
+use Magento\InventoryReservationsApi\Model\ReservationBuilderInterface;
 use Magento\TestFramework\Helper\Bootstrap;
 use PHPUnit\Framework\TestCase;
